test_script.py
import pytest
import pandas as pd
import json
from script import calculate_egfr, load_demographics, load_cmp_data, screen_patients_for_ckd

# Test data paths
demographics_file = 'test_data/sample_demographics.csv'
cmp_file = 'test_data/sample_cmp.json'

# Test eGFR calculation
def test_calculate_egfr():
    # Male example
    row_male = {'age': 45, 'sex': 'male', 'Serum_Creatinine': 1.2}
    egfr_male = calculate_egfr(row_male)
    assert round(egfr_male, 2) == 76.00
    
    # Female example (adjust expected result)
    row_female = {'age': 50, 'sex': 'female', 'Serum_Creatinine': 0.8}
    egfr_female = calculate_egfr(row_female)
    assert round(egfr_female, 2) == 89.71  # Adjusted expected result

# Test loading demographics data
def test_load_demographics():
    df_demographics = load_demographics(demographics_file)
    assert len(df_demographics) == 2
    assert df_demographics.loc[0, 'Patient_Name'] == 'John Doe'
    assert df_demographics.loc[1, 'mobile_number'] == '555-5678'

# Test loading CMP data
def test_load_cmp_data():
    df_cmp = load_cmp_data(cmp_file)
    assert len(df_cmp) == 2
    assert df_cmp.loc[0, 'Serum_Creatinine'] == 1.2
    assert df_cmp.loc[1, 'Serum_Creatinine'] == 0.8

# Test the merge function
def test_merge_data():
    df_demographics = load_demographics(demographics_file)
    df_cmp = load_cmp_data(cmp_file)
    
    df_demographics['patient_id'] = df_demographics['patient_id'].astype(str)
    df_cmp['Patient_ID'] = df_cmp['Patient_ID'].astype(str)
    
    df_merged = pd.merge(df_demographics, df_cmp, left_on='patient_id', right_on='Patient_ID')
    assert len(df_merged) == 2
    assert 'Patient_Name' in df_merged.columns

# Test filtering patients by eGFR <= 65
def test_filter_by_egfr():
    df_demographics = load_demographics(demographics_file)
    df_cmp = load_cmp_data(cmp_file)
    
    df_demographics['patient_id'] = df_demographics['patient_id'].astype(str)
    df_cmp['Patient_ID'] = df_cmp['Patient_ID'].astype(str)
    
    df_merged = pd.merge(df_demographics, df_cmp, left_on='patient_id', right_on='Patient_ID')
    
    df_merged['eGFR'] = df_merged.apply(calculate_egfr, axis=1)
    df_screened = df_merged[df_merged['eGFR'] <= 65]
    
    # Now expect 1 patient to be flagged (John Doe)
    assert len(df_screened) == 1

# Test complete CKD screening function
def test_screen_patients_for_ckd(tmpdir):
    output_file = tmpdir.join('results.csv')
    screen_patients_for_ckd(demographics_file, cmp_file, str(output_file))
    
    df_output = pd.read_csv(output_file)
    
    # Expect 1 patient to be flagged
    assert len(df_output) == 1
    assert df_output.loc[0, 'Patient_ID'] == 123
